Split legacy data helpers and rename active config utility module
- move legacy-only data helpers into legacy/utils/legacy_data_utils.py
- keep active src utility focused on config validation as src/utils/config_utils.py
- update legacy, tests, and stage0 imports to new module paths
- remove obsolete src/utils/data_utils.py
