+++ legacy/pre_schema_pipeline/src/get_seed_capability_results.py
@@ -8,8 +8,12 @@ import hydra
 import numpy as np
 from omegaconf import DictConfig
 
-from legacy.utils.legacy_data_utils import write_json_file
-from src.utils.data_utils import copy_file, list_dir, read_json_file
+from legacy.utils.legacy_data_utils import (
+    copy_file,
+    list_dir,
+    read_json_file,
+    write_json_file,
+)
 
 
 logger = logging.getLogger(__name__)
