+++ legacy/pre_schema_pipeline/src/run_evaluation.py
@@ -14,8 +14,7 @@ from legacy.utils import legacy_constants as constants
 from legacy.utils.legacy_capability_management_utils import (
     get_previous_capabilities,
 )
-from legacy.utils.legacy_data_utils import get_run_id
-from src.utils.data_utils import check_cfg
+from legacy.utils.legacy_data_utils import check_cfg, get_run_id
 
 
 @hydra.main(version_base=None, config_path="cfg", config_name="run_cfg")
